refactor(worker): split the REL-10 files under the repo size cap (REL10-R1)
`scripts/check_max_lines.py` (PROD-SIZE-1, 600 lines) failed on six files this
change grew. Founder-authorized scope amendment: split along real seams, with no
ratchets.toml entry and no behaviour change.

Rust
- new `test_support.rs` holds the typed trace collector and the instrumented
  listener that were duplicated between the two worker test modules;
- `cloud_client/mod.rs` keeps production code, its wire tests move to
  `cloud_client/tests.rs`;
- the REL-10 gate/fuse/403 cases move to `model_snapshot_sync_gate_tests.rs`.

Python
- new `tests/helpers/worker_heartbeat.py` holds the enrollment and heartbeat
  helpers the suites share;
- each of the three integration files hands its REL-10 class to a sibling file.

Test counts are unchanged: 92 Rust, 87 Python.
