+++ anyharness/crates/proliferate-worker/src/runtime_tests.rs
@@ -18,18 +18,16 @@
 //! current message, so renaming or adding an "expected skip" warning cannot slip
 //! past the silence assertions.
 
-use std::io::{Read, Write};
-use std::net::{SocketAddr, TcpListener};
-use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
-use std::sync::{Arc, Mutex};
-use std::time::Duration;
+use std::net::SocketAddr;
+use std::sync::atomic::{AtomicU64, Ordering};
 
-use tracing::field::{Field, Visit};
 use tracing::Level;
-use tracing_subscriber::layer::{Context, Layer};
-use tracing_subscriber::prelude::*;
 
 use super::{heartbeat_and_converge, TickControl};
+use crate::test_support::{
+    capture, events, is_request_containing, position, timeline, CapturedEvent, FieldValue,
+    InstrumentedServer, Route, Step, Timeline,
+};
 use crate::{
     cloud_client::CloudClient,
     config::WorkerConfig,
@@ -38,248 +36,6 @@ use crate::{
     observability::HEARTBEAT_ACK_TARGET,
     store::WorkerStore,
 };
-
-// ── Trace capture, typed ───────────────────────────────────────────────────
-
-/// Field values are kept TYPED: an acknowledgement that logged the verdict as
-/// the string `"false"`, or as a hard-coded constant of the wrong type, must not
-/// satisfy an assertion written for a real boolean.
-#[derive(Debug, Clone, PartialEq)]
-#[allow(dead_code)]
-enum FieldValue {
-    Bool(bool),
-    I64(i64),
-    U64(u64),
-    Str(String),
-    Debug(String),
-}
-
-impl FieldValue {
-    fn text(&self) -> String {
-        match self {
-            FieldValue::Bool(value) => value.to_string(),
-            FieldValue::I64(value) => value.to_string(),
-            FieldValue::U64(value) => value.to_string(),
-            FieldValue::Str(value) => value.clone(),
-            FieldValue::Debug(value) => value.clone(),
-        }
-    }
-}
-
-#[derive(Debug, Clone)]
-#[allow(dead_code)]
-struct CapturedEvent {
-    level: Level,
-    target: String,
-    fields: Vec<(String, FieldValue)>,
-}
-
-#[allow(dead_code)]
-impl CapturedEvent {
-    fn message(&self) -> String {
-        self.field("message")
-            .map(FieldValue::text)
-            .unwrap_or_default()
-    }
-
-    fn field(&self, name: &str) -> Option<&FieldValue> {
-        self.fields
-            .iter()
-            .find(|(key, _)| key == name)
-            .map(|(_, value)| value)
-    }
-}
-
-/// One ordered timeline shared by the collector and both listeners. Ordering
-/// claims are read off this single sequence rather than reconstructed by
-/// comparing separate request tallies against separate log dumps.
-#[derive(Debug, Clone)]
-#[allow(dead_code)]
-enum Step {
-    Event(CapturedEvent),
-    Request { server: &'static str, line: String },
-}
-
-type Timeline = Arc<Mutex<Vec<Step>>>;
-
-fn timeline() -> Timeline {
-    Arc::new(Mutex::new(Vec::new()))
-}
-
-fn steps(timeline: &Timeline) -> Vec<Step> {
-    timeline.lock().unwrap().clone()
-}
-
-fn events(timeline: &Timeline) -> Vec<CapturedEvent> {
-    steps(timeline)
-        .into_iter()
-        .filter_map(|step| match step {
-            Step::Event(event) => Some(event),
-            Step::Request { .. } => None,
-        })
-        .collect()
-}
-
-/// The index in the shared timeline of the first step matching a predicate.
-fn position(timeline: &Timeline, mut matches: impl FnMut(&Step) -> bool) -> Option<usize> {
-    steps(timeline).iter().position(|step| matches(step))
-}
-
-fn is_request_containing(needle: &str) -> impl Fn(&Step) -> bool + '_ {
-    move |step| matches!(step, Step::Request { line, .. } if line.contains(needle))
-}
-
-struct FieldCollector(Vec<(String, FieldValue)>);
-
-impl Visit for FieldCollector {
-    fn record_bool(&mut self, field: &Field, value: bool) {
-        self.0
-            .push((field.name().to_string(), FieldValue::Bool(value)));
-    }
-    fn record_i64(&mut self, field: &Field, value: i64) {
-        self.0.push((field.name().to_string(), FieldValue::I64(value)));
-    }
-    fn record_u64(&mut self, field: &Field, value: u64) {
-        self.0.push((field.name().to_string(), FieldValue::U64(value)));
-    }
-    fn record_str(&mut self, field: &Field, value: &str) {
-        self.0
-            .push((field.name().to_string(), FieldValue::Str(value.to_string())));
-    }
-    fn record_debug(&mut self, field: &Field, value: &dyn std::fmt::Debug) {
-        self.0.push((
-            field.name().to_string(),
-            FieldValue::Debug(format!("{value:?}")),
-        ));
-    }
-}
-
-struct CollectingLayer(Timeline);
-
-impl<S: tracing::Subscriber> Layer<S> for CollectingLayer {
-    fn on_event(&self, event: &tracing::Event<'_>, _context: Context<'_, S>) {
-        let mut collector = FieldCollector(Vec::new());
-        event.record(&mut collector);
-        let metadata = event.metadata();
-        self.0.lock().unwrap().push(Step::Event(CapturedEvent {
-            level: *metadata.level(),
-            target: metadata.target().to_string(),
-            fields: collector.0,
-        }));
-    }
-}
-
-fn capture(timeline: &Timeline) -> tracing::subscriber::DefaultGuard {
-    tracing::subscriber::set_default(
-        tracing_subscriber::registry().with(CollectingLayer(timeline.clone())),
-    )
-}
-
-// ── An instrumented listener that proves an ABSENT request ─────────────────
-
-/// `(method, exact path, status, body)`.
-type Route = (&'static str, String, u16, String);
-
-/// A fake HTTP server that answers a routing table for as long as the test wants
-/// and records every request. It never pre-commits to a request count, which is
-/// exactly what makes "zero snapshot requests arrived" an assertion rather than
-/// a hang.
-#[allow(dead_code)]
-struct InstrumentedServer {
-    address: SocketAddr,
-    hits: Arc<Mutex<Vec<String>>>,
-    stop: Arc<AtomicBool>,
-    handle: Option<std::thread::JoinHandle<()>>,
-}
-
-#[allow(dead_code)]
-impl InstrumentedServer {
-    fn spawn(label: &'static str, routes: Vec<Route>, timeline: Timeline) -> Self {
-        let listener = TcpListener::bind("127.0.0.1:0").expect("bind instrumented server");
-        listener
-            .set_nonblocking(true)
-            .expect("non-blocking instrumented listener");
-        let address = listener.local_addr().expect("instrumented server address");
-        let hits: Arc<Mutex<Vec<String>>> = Arc::new(Mutex::new(Vec::new()));
-        let thread_hits = Arc::clone(&hits);
-        let stop = Arc::new(AtomicBool::new(false));
-        let thread_stop = Arc::clone(&stop);
-        let handle = std::thread::spawn(move || {
-            // A polled accept loop, never a blocking one: the only way a test
-            // can prove "no request arrived" is if stopping the server never
-            // depends on a request arriving.
-            loop {
-                if thread_stop.load(Ordering::SeqCst) {
-                    break;
-                }
-                let mut stream = match listener.accept() {
-                    Ok((stream, _)) => stream,
-                    Err(error) if error.kind() == std::io::ErrorKind::WouldBlock => {
-                        std::thread::sleep(Duration::from_millis(2));
-                        continue;
-                    }
-                    Err(_) => break,
-                };
-                stream
-                    .set_nonblocking(false)
-                    .expect("blocking accepted stream");
-                stream
-                    .set_read_timeout(Some(Duration::from_secs(5)))
-                    .expect("accepted stream read timeout");
-                let mut buffer = vec![0u8; 262_144];
-                let read = match stream.read(&mut buffer) {
-                    Ok(0) => continue,
-                    Ok(read) => read,
-                    Err(_) => continue,
-                };
-                let raw = String::from_utf8_lossy(&buffer[..read]).to_string();
-                let (method, path) = request_line(&raw);
-                let line = format!("{method} {path}");
-                thread_hits.lock().unwrap().push(line.clone());
-                timeline
-                    .lock()
-                    .unwrap()
-                    .push(Step::Request { server: label, line });
-                let (status, body) = routes
-                    .iter()
-                    .find(|(route_method, route_path, _, _)| {
-                        *route_method == method && *route_path == path
-                    })
-                    .map(|(_, _, status, body)| (*status, body.clone()))
-                    .unwrap_or((404, "{}".to_string()));
-                let _ = write!(
-                    stream,
-                    "HTTP/1.1 {status} X\r\nContent-Type: application/json\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{body}",
-                    body.len()
-                );
-            }
-        });
-        Self {
-            address,
-            hits,
-            stop,
-            handle: Some(handle),
-        }
-    }
-
-    /// Stop the accept loop and drain the thread, so a test's request tally is
-    /// final rather than racing a late connection.
-    fn shutdown(mut self) -> Vec<String> {
-        self.stop.store(true, Ordering::SeqCst);
-        if let Some(handle) = self.handle.take() {
-            handle.join().expect("instrumented server thread");
-        }
-        self.hits.lock().unwrap().clone()
-    }
-}
-
-fn request_line(raw: &str) -> (String, String) {
-    let mut parts = raw.lines().next().unwrap_or_default().split(' ');
-    (
-        parts.next().unwrap_or_default().to_string(),
-        parts.next().unwrap_or_default().to_string(),
-    )
-}
 
 // ── Fixtures ───────────────────────────────────────────────────────────────
